Extract keeps signers data to separate package
Extracted keeps and signers initialization to a pakcage containing test
data to use it across tests.
